further refactoring to make Time computations nicer
