+++ community/values/src/main/java/org/neo4j/values/storable/DurationValue.java
@@ -55,10 +55,10 @@ import static java.util.Objects.requireNonNull;
 import static java.util.regex.Pattern.CASE_INSENSITIVE;
 import static org.neo4j.values.storable.NumberType.NO_NUMBER;
 import static org.neo4j.values.storable.NumberValue.safeCastFloatingPoint;
-import static org.neo4j.values.storable.TimeConstants.AVG_DAYS_PER_MONTH;
-import static org.neo4j.values.storable.TimeConstants.AVG_SECONDS_PER_MONTH;
-import static org.neo4j.values.storable.TimeConstants.NANOS_PER_SECOND;
-import static org.neo4j.values.storable.TimeConstants.SECONDS_PER_DAY;
+import static org.neo4j.values.storable.TimeUtil.AVG_DAYS_PER_MONTH;
+import static org.neo4j.values.storable.TimeUtil.AVG_SECONDS_PER_MONTH;
+import static org.neo4j.values.storable.TimeUtil.NANOS_PER_SECOND;
+import static org.neo4j.values.storable.TimeUtil.SECONDS_PER_DAY;
 
 /**
  * We use our own implementation because neither {@link java.time.Duration} nor {@link java.time.Period} fits our needs.
